Refactoring map overlay components
