+++ src/components/common/map-overlays/MapOverlays.js
@@ -1,13 +1,10 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import { Polyline } from 'react-native-maps';
 
 import AircraftMarker from './AircraftMarker';
 import ControllerPolygon from './ControllerPolygon';
 
-import colors from '../../../config/colors.json';
-
-export const MapOverlays = ({ clients, focusedClient, setFocusedClient }) => (
+const MapOverlays = ({ clients, focusedClient, setFocusedClient }) => (
   <>
     {clients.map(client => {
       const isFocusedClient = focusedClient.callsign === client.callsign;
@@ -36,40 +33,10 @@ export const MapOverlays = ({ clients, focusedClient, setFocusedClient }) => (
   </>
 );
 
-export const FlightPath = ({ client: { depCoords, location, arrCoords } }) => {
-  // Render polylines only if airport coords are present
-  if (depCoords && location && arrCoords) {
-    const { lineFlown, lineRemaining } = colors.mapOverlays;
-    return (
-      <>
-        <Polyline
-          coordinates={[depCoords, location]}
-          strokeColor={lineFlown}
-          strokeWidth={2}
-          zIndex={5}
-          geodesic
-        />
-        <Polyline
-          coordinates={[location, arrCoords]}
-          strokeColor={lineRemaining}
-          strokeWidth={2}
-          zIndex={5}
-          geodesic
-        />
-      </>
-    );
-  }
-
-  // Render nothing if polylines can't be rendered
-  return null;
-};
-
 MapOverlays.propTypes = {
   clients: PropTypes.array.isRequired,
   focusedClient: PropTypes.object.isRequired,
   setFocusedClient: PropTypes.func.isRequired,
 };
 
-FlightPath.propTypes = {
-  client: PropTypes.object.isRequired,
-};
+export default MapOverlays;
